Move typecheck-specific code to rsc.typecheck
Very soon, Rsc will gain the capability to load SemanticDBs produced
by metacp and use that information to understand third-party dependencies.

In preparation for that, I asked myself a question: "Which parts of
rsc.semantics make sense both for metadata loaded from SemanticDBs
and metadata computed during typechecking?".

Answering that question led me to believe that lots of data structures
and logic in rsc.semantics belongs to rsc.typecheck instead.
For example, the ceremony around scope statuses, the Scope.lookup/resolve
separation - all that is irrelevant to SemanticDB loading (at least,
as far as I can see now).
