+++ bench/rsc/jvm/src/main/scala/rsc/bench/RscOutline.scala
@@ -6,7 +6,6 @@ import java.util.concurrent.TimeUnit
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.annotations.Mode._
 import rsc.bench.RscOutline._
-import rsc.semantics._
 
 object RscOutline {
   @State(Scope.Benchmark)
