+++ bench/rsc/jvm/src/main/scala/rsc/bench/RscScope.scala
@@ -6,7 +6,6 @@ import java.util.concurrent.TimeUnit
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.annotations.Mode._
 import rsc.bench.RscScope._
-import rsc.semantics._
 
 object RscScope {
   @State(Scope.Benchmark)
