+++ rsc/src/main/scala/rsc/pretty/PrettyStatus.scala
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
 package rsc.pretty
 
-import rsc.semantics._
+import rsc.typecheck._
 
 object PrettyStatus {
   def str(p: Printer, x: Status): Unit = {
