+++ rsc/src/main/scala/rsc/typecheck/Envs.scala
@@ -1,9 +1,10 @@
 // Copyright (c) 2017-2018 Twitter, Inc.
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
-package rsc.semantics
+package rsc.typecheck
 
 import scala.annotation.tailrec
 import rsc.pretty._
+import rsc.semantics._
 import rsc.util._
 
 sealed class Env protected (val _scopes: List[Scope]) extends Pretty {
