+++ rsc/src/main/scala/rsc/typecheck/Resolutions.scala
@@ -1,8 +1,9 @@
 // Copyright (c) 2017-2018 Twitter, Inc.
 // Licensed under the Apache License, Version 2.0 (see LICENSE.md).
-package rsc.semantics
+package rsc.typecheck
 
 import rsc.pretty._
+import rsc.semantics._
 
 sealed trait Resolution extends Pretty with Product {
   override def printStr(p: Printer): Unit = PrettyResolution.str(p, this)
